Switch to mvvm::int64 instead of `long long`

+++ src/lib/sup-mvvm-viewmodel/mvvm/editors/editor_builders.cpp
@@ -112,7 +112,7 @@ editorbuilder_t LongIntSpinBoxEditorBuilder()
     {
       if (item->HasData(DataRole::kLimits))
       {
-        auto limits = item->Data<Limits<long long>>(DataRole::kLimits);
+        auto limits = item->Data<Limits<int64>>(DataRole::kLimits);
         editor->SetRange(limits.GetLowerLimit(), limits.GetUpperLimit());
       }
     }
